Add People app for tracking authors
The developer API is going to become intolerably annoying if Authors
aren't objects in the database.

Down the line, we'll also need Liaisons to exist, and they would
naturally live in the people app. DLCs will probably want to be
separated out at that point, and they will probably belong here as well.