Add Liaison and DLC as database objects
We're going to need them to be separate objects in order to support
well-formed queries and do recordkeeping.

This commit requires blowing away your existing database; it recreates
all the migrations from scratch. It also updates records.yaml so the
database can be repopulated with the earlier objects in the new schema.