The author's name is set correctly when we import a record
This is a big commit because it required building all of the CSV
parsing machinery, which was weirdly complicated (see
`records.views.Import._get_csv_reader()`). It also necessitated a
migration on DLC - we can't require Liaison because we don't have that
information when we import records. That'll be OK from a workflow
perspective, though, since we already expect to need to manually set
liaisons.