Make sure we're storing the MIT ID hash, not the MIT ID itself.
The ID is sensitive data and we're not allowed to use it on offsite
servers. However, we don't need to actually know it - we just need to
be able to track author uniqueness and tell whether or not we already
have a given author in our database. A hash of the ID will serve this
purpose just fine.