Remove grunt dependencies
It turns out we don't need this! It worked when we originally deployed
it to Heroku *because the nodejs buildpack ran before the python
buildpack*, meaning that the compressed assets existed when whitenoise
checked for static files on boot.

However, we don't need grunt to compile static assets - we can do that
entirely within the python buildpack, using libsass - which is simpler,
hence cleaner. This in turn, however, is why we needed to add the
post_compile hook in the previous commit - to trigger the `python
manage.py compress` command prior to the point where whitenoise loads
static.