Perform radical surgery to remove record.status [Fixes #31]
Record is the wrong place to keep status:
* we're aggressively rejecting INVALID records
* SENT belongs properly on EmailMessage

So this commit removes anything pertaining to INVALID; adds an FK
from Record to EmailMessage; removes status and status_timestamp from
Record; and cleans up whatever broke as a result.

This should also allow for mostly(*) easier EmailMessage construction,
since an EmailMessage really is a container for a set of Records.

(*) Mostly because users would like to edit the text of emails
_including the citation area_, and it's a little unclear how we handle
updates in that case.