Handle multiple records for the same paper with different authors
It's OK to have multiple records with the same PaperID if they have
different authors. However, PaperID and Author must be unique_together.

Add this constraint; look for and resolve places we assumed PaperID was
unique; reject records when we've already emailed a different author
for the paper; add test coverage.

Resolves #37.