Experiment with asynchronous gunicorn workers
Timeouts to the Elements API call are causing the user-facing part of
the system to have application errors. This shouldn't happen - the
whole point of using signals was to keep the API http request out of the
user-facing request/response process - but the default synchronous
gunicorn workers can block, so let's try this.