Make selected citations more prominent
Users expressed concern that it may be hard to see status on a long
list of citations. We don't want to show/hide citations at this stage
because that may lead to confusing interactions; instead, we will
highlight selected to make status easier to see, and await user
reactions.

Fixes #58.