Make a warning show when the dspace url in the email scrolls into view
Sometimes it doesn't work because the author name form in dspace may
not match my available name form, and I have no way to check. I don't
want to have a warning that shows all the time because people will
ignore it. Maybe this helps?

Also, figuring out when I could actually bind to the elements, and
how to find elements inside an iframe, was brutal.