Add utf-16le to the supported encodings
If you export from Tableau as crosstab, you get a tab-delimited,
utf-16 encoded file with BOM. Because it's an InMemoryUploadedFile, not
just a thing on the filesystem, you can't just close and reopen it with
the desired encoding (which would be easy and Just Work); there's a
song-and-dance to make it readable. And because of the tab delimiter,
you have to specify a dialect or it doesn't properly split the data.
Fun times.