Look for places that rogue connections might be left open
We are still having an intermittent problem of too many db connections
being open, and we don't have enough logging to track it down. Working
on the logging question separately, but for now I'm looking for
processes that might have opened but not closed connections, and adding
some cleanup. The Import view is a particular worry here as it is
likely to crash with a Heroku timeout error, thereby preventing Django
from doing its usual cleanup. The LoginRequiredMixin shouldn't be a
problem - it's squarely in the request-response loop and a thin wrapper
around Django's native decorator - but we're going to hit it essentially
every time users do anything (and it's a place we have hit an
OperationalError in the wild, opening a connection to check if the user
is authenticated), so we may as well close old connections here.