Make logging out appear to work
Clicking the logout button does in fact flush the user. However, when
the user clicks links to other pages, the login requirement checks
with social auth, which finds an active token and re-logs them in.
The user doesn't see this (although there are traces in the logs), so
it appears to the user as if they have not been logged out.

python-social-auth naturally provides a *disconnection* method, where
you sever the social and django accounts. We don't want that, though,
since we want people to be able to return to their accounts via OAuth.
Instead, we just want the tokens to be revoked so that users aren't
automatically logged back in.

python-social-auth wraps its functions in a lot of assumptions about
their context, which makes it extremely difficult to pull them out and
use them individually (figuring out how to supply working arguments is
hard). So the easiest thing to do, given that we don't expect to want
to actually disconnect users, is to use the built-in disconnection
function, but shorten the pipeline so it only runs token revocation.

Then we wrap the Django LogoutView in a local one that also runs
the disconnection pipeline. Voila!