Change time window for reporting management command
The most recent report indicated that 44 calls had timed out, when in
fact they were issued successfully, but reporting (it appears) ran
between the time the calls were initiated and the time that the http
responses were received. Offsetting the reporting window a bit
backward from the current time will ensure that we are not
examining any calls in progress.