Return response text in patch elements call
Why these changes are being introduced:

The patch elements function was returning the response object from
Elements, but task values now need to be json-serializable, so this was
raising an error. The task was still functioning correctly, it just
raised the error on the return value.

How this addresses that need:
Changes the return value to the response text instead of the response
object, since the text is json-serializable.

Side effects of this change:
None

Relevant ticket(s):
NA