Extend character limit on record model fields
Why these changes are being introduced:
There was previously an arbitrarily short character limit of 45 on the
Record model doi CharField. DOIs do not have a character limit
restriction, and we did in fact encounter errors when attempting to
import papers with DOIs longer than 45 characters.

How this addresses that need:
* Extends character limit on Record.doi field to 255 characters
* Also extends the (similarly arbitrarily short) character limit on a
  few other fields in the Record model

Side effects of this change:
None, but a note that tests have not been updated because SQLite (which
is used for the test database) does not enforce the max_length property
on VARCHAR fields, so adding a test for this would be pointless as it
will pass regardless of DOI length or max_length setting.